app: aboot: Ability to flash only one volume for UBIFS partitions
UBIFS partitions may have multiple UBI volumes written to them. For example:
system partition image may hold: system, rootfs, userdata and cache volumes.
With current implementation the only way to update just one of the volumes is to
flash the entire system image partition.
Note that currently only system image is considered to hold multiple volumes.
The provided volume image should be a UBIFS image.

This patch adds the ability to update the data for just one volume.
The command used is:
fastboot flash <volume name> <volume image - UBIFS>
For example:
fastboot flash usrfs usrfs.ubifs
Due to the difficulty in adding a new fastboot command the existing "fastboot flash"
command implementation was reused:
When trying to flash a partition, if the provided partition name does not correspond
with any partition, treat the call as a ubiupdatevol. Meaning, search for a
corresponding volume.
The above is working due to the fact that volumes and partitions are named
differently.

Note: It was cherry-pick from 84eaaf6ebe9d274c64d50105933e41b2b021f95a

Resolves: QTI9X40-1851
Change-Id: I9f2ea580c3a07955925c062df1302bf3194f27b9
Signed-off-by: Caupar Gu <CGu@sierrawireless.com>