aboot: Fix data abort on "fastboot boot <img>"
This is a temporary fix for Qualcomm case 02205852: data abort
when using "fastboot boot <image>".

The reason for this data abort was that as part of copying DTB,
kernel tags were copied to location 0x88000000 for mdm9x40, which
overlaps with the start of downloaded image (coincidentially at
0x88000000). As a result, the image header was overwritten and a
faulty kernel and ramdisk address and size were passed to loader.
The crash hence occurred when attempting to copy the ramdisk.

The fix is to copy the downloaded header into a local one and use
the local header subsequently. Also copy the DTB after copying the
kernel and ramdisk to prevent corrupting the two images.

This is a temporary fix until Qualcomm comes up with a recommended
solution (if at all).

Change-Id: Ic3ac0298dc356fc3c9b8acee85e05636d7a340f6
Signed-off-by: Zoran Markovic <zmarkovic@sierrawireless.com>