lk: [9x28] Switch 4k page image to 2k for UBI with multi volumes
Description:
There is an function in LK, when the default image being to burn
to the flash device is UBI format and this image is for 4K page only,
then it will auto detect and change the image to 2K page size before
burning to the flash device. But this function have some issue:
1,It can only change the UBI image that only contain one volume in it,
we should change the function to support that it can change the UBI
image to 2k page size with multi volumes in it; 2,This function can
not correctly change the static UBI volume.

Notes: This change is merge from QTI9X40-545
Solution:
Change the code to support multi volume or static volume image switch
from 4k page to 2k page.

Resolves: QTI9X28-2653
Signed-off-by: Caupar Gu <CGu@sierrawireless.com>
Change-Id: Idd63676a88a9033f7fea7c0ba13d5f79ab708959