dev: gcdb: display: check for valid DFPS base address
When an APPS bootloader image is flashed on device,  it is
possible that the dfps structure in LK is different from the one
that is stored in splash partition due to mismatch in value of
DFPS_MAX_FRAME_RATE. Check for validity of dfps base address that
is copied from splash partition with the expected address in LK.
If both are not equal, re-calculate the DSI PLL trim codes and
store it back to splash partition. This would avoid issues due
to access of wrong dfps base address and incorrect trim codes.

Change-Id: I19e1315baae408fe6864191daacb4ac526afaf7c