Add the "BOOTQUIETDISABLE" NV and AT command
Issue Description:
Add a NV and AT command to control whether show the
kernel boot message.

Root Cause:
N/A

Solution:
Add a bcbootquiet byte to bccoworkmsg struct for
"BOOTQUIETDISABLE", and Read it to update cmdline.
When the "BOOTQUIETDISABLE" was set 0, the commandline
will have a "quiet" paramter, the kernel bootup message
will be hide. When the "BOOTQUIETDISABLE" was set 1, the
kernel bootup message will be show.

Resolves: QTI9X28-605
Change-Id: I71c1f50982a55daccfe00edb125d4083e1527c97
Signed-off-by: Lucy He <LHe@sierrawireless.com>