Improve the EFS restore when system swap for badimage reason
Description:
Improve the EFS restore when system swap for badimage reason. When the 
system swap-reason flag is not up to date. A EFS restore will be triggered 
unexpectly.

Root Cause:
In current design, the trigger of  efs-restore for badimage rely on the 
swap-reason flag in DS smem.

Effect Platform:
9x28/9x40

Solution:
It is more resonable to trigger the efs-restore when swap reason is for 
badimage.

Resolves: QTI9X40-1579
Change-Id: I0d4877234f22a6b5a3265f567d6dab57ac58dcc2
Signed-off-by: Felix Zhu <FZhu@sierrawireless.com>