New action of continual multi abnormal resets when at!eroption = 0
Investigation:
With the latest FW, it will not go into dump mode but swap system when 6 abnormal reset with at!eroption is set to dump mode.

Effect Platform:
9x28/9x40.

Root cause:
When errorcount is 3, it will delete efs fs. And in next abnormal resets, the efs hadn't restored, so in sbl the eroption is reset, so it will swap system when errorcount is 6.

Solution:
When at!eroption is set to dump mode. When errorcount is 3, backup eroption to smem. After 5 abnormal reset, module should go into dump mode, then after catch dump, it will enter recovery mode.

Resolves: QTI9X28-1458
Change-Id: Ic387e51bc15cf1985076dcfe28590a70e5a4aba7
Signed-off-by: Leon Peng <lpeng@sierrawireless.com>