Add interface to get the data of a UBI volume from UBI partition
Descrition:
This patch adds interface to get all the data of a specified
UBI volume from a UBI partition.

Usage:
Create a buffer that is bigger then the UBI volume size, create
a pointer *size to store the data size. And then call function
"get_ubi_vol_data" with required parameters, In case of successful
all the volume data will save in that buffer and the file size is
save in pointer *size.

Resolves: QTI9X40-1851
Change-Id: Ibb7067c763cbd9c30773c4a16afd6e27bdc713a4
Signed-off-by: Caupar Gu <CGu@sierrawireless.com>