Allow resize UBI volume during UBI volume update
Description:
Currently when try to update one of the UBI volumes in UBI
partition, the new image size must not bigger than the
currently UBI volume reserved size, or it will lead to
update fail.

Solution:
Resize the UBI volume for new image when the reserved size
of the UBI partition is big enough.

Change-Id: I04de0077706d8e95714224ac2894e8757874da0b
Resolvers: QTI9X28-1629
Signed-off-by: Caupar Gu <CGu@sierrawireless.com>
Signed-off-by: Rock Li <RLi@sierrawireless.com>