Allow resize UBI volume during UBI volume update
Description:
Currently when try to update one of the UBI volumes in UBI
partition, the new image size must not bigger than the
currently UBI volume reserved size, or it will lead to
update fail.

Solution:
Resize the UBI volume for new image when the reserved size
of the UBI partition is big enough.

Resolvers: QTI9X40-1851
Change-Id: I4a5bd085cf5c343bbbc27a2e146f440a3e557d96
Signed-off-by: Caupar Gu <CGu@sierrawireless.com>