lk: Module freeze on fastboot mode, fail to boot up
Issue Descrition:
Module freeze on fastboot mode, failed to reboot mdoule.

Investigation:
1, Reproduce this issue, find device state working normal,
   except can't reset by fastboot reset command.
2. Measure the PS_HOLD pin, it has been pull low.
3. If press reset key, device can reset.
4. check the reset code of LK, it do below process
 a. Disable PS_HOLD reset (Write PMIC register)
 b. Set reset type (Write PMIC register)
 c. Re-enable PS_HOLD reset.
 d. assert PS_HOLD pin to low
5. Add debug log to check the rest-type is set correct or not.
6. Test device 4W times, not reproduce it.

Root Cause:
From the code and test result, it seems due to the timing
of writing PMIC registers.

Solution:
Add the debug log as a workaround of this issue.
And open another action to trace the test state.

Change-Id: Ibaf988151835db17e68f013cea3d1921b8a45f02
Resolves: QTI9X28-2297
Signed-off-by: Traving Chen <trchen@sierrawireless.com>