ima: read fuse to force IMA to be enabled
If the IMA fuse bit is set, the IMA will be forced to enabled.
If else, both LK (one with IMA enabled, one without) are provided
for development purpose.
If LK detects the IMA fuse is set but the build is NOT IMA ready,
LK rejects the load of the kernel and performs a swap on dual
systems or enters to recovery mode for single systems.

note: modifications to commit for dual system stubs in wp

Resolves: LXSWIREF-746
Related: LXSWIREF-543
Related: QTI9X28-3727
Change-Id: I9fa674c12521c6b7279391dfe0a78a826c70530a
Signed-off-by: Christophe Gottheimer <cgottheimer@sierrawireless.com>
Modified-by: Steve MacDonald <smacdonald@sierrawireless.com>
Signed-off-by: Steve MacDonald <smacdonald@sierrawireless.com>