ima: Check if kernel is IMA ready before loading a kernel
If the eFuse ima_flag is set to 1, or if IMA is expected to be
enabled, check if kernel is IMA ready. Else, reject the download
and swap to older system (Dual Systems targets) or enter into
emergency download (Single System targets).

note: modifications to commit for dual system stubs in wp

Resolves: LXSWIREF-746
Related: LXSWIREF-574
Change-Id: I78bcf6d03f75d4ef12e9a7ad6058dcff411494ac
Signed-off-by: Christophe Gottheimer <cgottheimer@sierrawireless.com>
Modified-by: Steve MacDonald <smacdonald@sierrawireless.com>
Signed-off-by: Steve MacDonald <smacdonald@sierrawireless.com>