ima: LK enters bootloader when kernel is not IMA ready
If the kernel is not detected as IMA ready, the key "ima_ready=1"
is missing on the command line, to ensure backward compatibility,
LK will enter into bootloader to give a chance to flash a good
image. This is only possible if the eFuse IMA is not burnt.

Resolves: LXSWIREF-746
Related: LXSWIREF-680
Change-Id: I8d0a03c3e178100f706b06624c0cefff32107f41
Signed-off-by: Steve MacDonald <smacdonald@sierrawireless.com>