reliability: Stack overflow detected in LK
Description:
During pen-testing activities, a stack overflow vulnerability
has been detected in the Little Kernel (APPSBL).When parsing
the SSData partition in the sierra_ds_dssd_partition_read function,
a call is made to sierra_ds_dssd_data_blocks_find. This function
returns an array of legal blocks found in the partition. The
pointer sent to the function to contain this array is a stack
pointer. There is no verification that the partition does not
contain less than the size of the sent stack buffer. By controlling
this partition, it may cause a stack overflow.

Solution:
Extend the array data_block[] from 2 to 24, and do data block check in
sierra_ds_dssd_data_blocks_find()

Resolves: QTI9X40-6257
Change-Id: I0f4a0ee430dafe9f4ca4551a17e6352ce0e8ceeb
Signed-off-by: lchen <lchen@sierrawireless.com>